// src/components/providers/pragmaticplay/enums/callback_methods.enum.ts
export enum CallbackMethods {
    Authenticate = 'authenticate.html',
    Balance = 'balance.html',
    Bet = 'bet.html',
    Result = 'result.html',
    BonusWin = 'bonusWin.html',
    JackpotWin = 'jackpotWin.html',
    EndRound = 'endRound.html',
    Refund = 'refund.html',
    Withdraw = 'notify/withdraw.html',
    GetBalancePerGame = 'getBalancePerGame.html',
    PromoWin = 'promoWin.html',
}